Create homepage header css 1st draft
I want to create a vault theme with the way my page is setup. In the
future I want an intro vault hissing and opening animation into the
homepage where the user will see a 3d rendition of the menu and content.
Something similar is Persona 5 Royals intro sequence to the load game
menu.

Humble origins though. I've decided to use the rich black background and
celadon green text for the whole page instead of switching it up with
other colors for consistency and aesthetics for now. I'll style it
differently once there's enough content.